requirements.txt
django==3.0.8
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
django-cors-headers==3.4.0
django-rest-knox==4.1.0
pillow==7.2.0